backend/ticketing_service/requirements.txt
asgiref==3.8.1
celery==5.4.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
injector==0.22.0
mysqlclient==2.2.4
packaging==24.1
PyJWT==1.7.1
python-dotenv==1.0.1
requests==2.32.3
redis==5.0.8
sqlparse==0.5.1
tzdata==2024.1
